+++ pyPDEs/mesh/__init__.py
@@ -1,4 +1,6 @@
-from .mesh import Mesh, Cell, Face
+from .mesh import Mesh
+from .cell import Cell
+from .face import Face
 
 from .create_mesh import create_1d_mesh
 from .create_mesh import create_2d_mesh
